Folder reorganization, added LevelDB libraries as binary...
... but only temporarily. In the future they should be build from scratch on "npm install".
