+++ .gitignore
@@ -1,16 +1,6 @@
 /.vscode/
-/_backups/
-/bin/
 /build/
 /node_modules/
-/src/build.bat
-/src/tw*
-src/render/blocks/
-/src/dev/
-/src/buffer.txt
-index.js
-/src/rp/
-/src/Benchmark.png
 npm-debug.log
 /dist/
 package-lock.json
