+++ db/LevelDbWrapper.js
