+++ db/trimChunk.js
